Make way for OSXBMC Python Changes.
You must download the latest experimental version of riquedafreak's OSXBMC fork.  This fork includes several new features for python that allows python scripts greater manipulation of the UI environment.

First on the list is breadcrumbs, we're trying to create dynamic breadcrumbs that are created each time you traverse a folder in the library.  Right now it is very rudimentary as Python does not have very much support for ControlGroup.