Simplify the skin.
Lets simplify the skin by going straight into the UI portion of the skin.  I was growing tired of watch the splash screen each time.

Also we added a fade-in / fade-out feature, that whenever a window is opened and closed it is faded-in and faded-out respectively.  This fade affect is done by putting a solid black image over the entire display and changing the opacity for 100% to 0% which causes an apparent fading-in of the window.  We reverse the process to get a fade-out affect.