Zoom on select.
Added cool animation, similar to the flash demo, that zooms the menu Icon when it is clicked on.  There are now two icons to represent the same menu item, one with and one without a shadow.  Maybe there should be 1 icon, and the reflection is just a flip of that original icon.